Adds gamelogic.js file and moved the RockPaperScissors function to it and adds script tag to html
