+++ spec.helper.js
@@ -2,4 +2,4 @@ const chai = require('chai');
 const BrowserHelpers = require('e2e_training_wheels')
 global.browser = new BrowserHelpers()
 global.expect = chai.expect;
-global.RockPaperScissors = require('./src/js/rockpaperscissors')
+global.RockPaperScissors = require('./src/js/gamelogic')
