+++ src/index.html
@@ -29,6 +29,7 @@
             </div>
 
         <script src="js/rockpaperscissors.js"></script>
+        <script src="js/gamelogic.js"></script>
         
     </body>
 </html>
