+++ src/js/gamelogic.js
@@ -1,0 +1,38 @@
+function RockPaperScissors(choiceOne) {
+  const choiceTwo = computerChoice();
+  
+  if(choiceOne === choiceTwo){
+  return result_div.innerHTML = "The result is a tie!"
+  };
+  
+  if(choiceOne === 'rock'){
+    if(choiceTwo === 'scissors'){
+    return win(choiceOne, choiceTwo)
+    }
+    else{
+      return lose(choiceOne, choiceTwo)
+    }
+  };
+
+  if(choiceOne === 'paper'){
+    if(choiceTwo === 'rock'){
+      return win(choiceOne, choiceTwo)
+    }
+    else{
+      return lose(choiceOne, choiceTwo)
+    }
+  };
+
+  if(choiceOne === 'scissors'){
+    if(choiceTwo === 'paper'){
+      return win(choiceOne, choiceTwo)
+    }
+    else{
+      return lose(choiceOne, choiceTwo)
+    }
+  }
+};
+
+if (typeof module !== 'undefined' && module.exports) {
+  module.exports = RockPaperScissors;
+}
