+++ src/js/rockpaperscissors.js
@@ -41,41 +41,3 @@ function lose(user, computer) {
   result_div.innerHTML = `${computer} beats ${user}. You lose!`
 }
 
-function RockPaperScissors(choiceOne) {
-  const choiceTwo = computerChoice();
-  
-  if(choiceOne === choiceTwo){
-  return result_div.innerHTML = "The result is a tie!"
-  };
-  
-  if(choiceOne === 'rock'){
-    if(choiceTwo === 'scissors'){
-    return win(choiceOne, choiceTwo)
-    }
-    else{
-      return lose(choiceOne, choiceTwo)
-    }
-  };
-
-  if(choiceOne === 'paper'){
-    if(choiceTwo === 'rock'){
-      return win(choiceOne, choiceTwo)
-    }
-    else{
-      return lose(choiceOne, choiceTwo)
-    }
-  };
-
-  if(choiceOne === 'scissors'){
-    if(choiceTwo === 'paper'){
-      return win(choiceOne, choiceTwo)
-    }
-    else{
-      return lose(choiceOne, choiceTwo)
-    }
-  }
-};
-
-if (typeof module !== 'undefined' && module.exports) {
-  module.exports = RockPaperScissors;
-}
